refactor: 사용자 시퀀스 타입 변경

+++ src/main/java/ringle/first/assignment/user/entity/User.java
@@ -19,7 +19,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 public class User {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userSeq;
+    private long userSeq;
 
     @NotNull
     private String userId;
